GABRIEL/TAREA_1/requirements.txt
 dnspython
 argparse